requirements.txt
annotated-types==0.7.0
biopython==1.85
blis==0.7.11
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
conllu==6.0.0
cymem==2.0.11
en_ner_bc5cdr_md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_bc5cdr_md-0.5.4.tar.gz#sha256=ffc73130a710edf851206199720cb2c744a043e032f5da6ba4bb36863deca778
en_ner_bionlp13cg_md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_bionlp13cg_md-0.5.4.tar.gz#sha256=d747ec2b3871d7f9187d4f4769e3f7014a48e804b01c18e3da9991a67024eae1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
nmslib-metabrainz==2.1.3
numpy==1.26.4
packaging==24.2
pandas==2.2.3
preshed==3.0.9
psutil==7.0.0
pybind11==2.13.6
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pysbd==0.3.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
scispacy==0.5.5
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.2.5
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
